Remove the option to change the table size
The options would let the user update the size of the graph in an embed. This is an issue because the height of an iframe can't be updated.

In addition, when the share modal would open, the number of rows per page wouldn't be restored because it's not part of the widget's state.

As a consequence, it's easier to remove this option and see if it's really needed.